BJS2-54228 Перенос логики в projectController и projectService, правка проверки на максимальное кол-во файлов для проекта
