+++ src/main/java/faang/school/projectservice/service/s3/S3Service.java
@@ -4,11 +4,7 @@ import com.amazonaws.SdkClientException;
 import com.amazonaws.services.s3.AmazonS3;
 import com.amazonaws.services.s3.model.ObjectMetadata;
 import com.amazonaws.services.s3.model.PutObjectRequest;
-import com.amazonaws.services.s3.model.S3ObjectSummary;
 import faang.school.projectservice.exception.IntegrationException;
-import faang.school.projectservice.model.Resource;
-import faang.school.projectservice.model.ResourceStatus;
-import faang.school.projectservice.model.ResourceType;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Value;
@@ -16,9 +12,6 @@ import org.springframework.stereotype.Service;
 import org.springframework.web.multipart.MultipartFile;
 
 import java.io.IOException;
-import java.math.BigInteger;
-import java.time.LocalDateTime;
-import java.util.List;
 
 @Service
 @Slf4j
@@ -30,25 +23,18 @@ public class S3Service {
 
     private final AmazonS3 s3Client;
 
-    public Resource uploadFile(MultipartFile file, String folder) {
+    public String uploadFile(MultipartFile file, String folder) {
         String key = String.format("%s/%d%s", folder, System.currentTimeMillis(), file.getOriginalFilename());
         putObjectToStorage(file, key);
 
-        return Resource.builder()
-                .key(key)
-                .name(file.getName())
-                .size(BigInteger.valueOf(file.getSize()))
-                .type(ResourceType.getResourceType(file.getContentType()))
-                .status(ResourceStatus.ACTIVE)
-                .createdAt(LocalDateTime.now())
-                .build();
+        return key;
     }
 
     public void deleteFile(String key) {
         try {
             s3Client.deleteObject(bucketName, key);
         } catch (SdkClientException exception) {
-            String errorMessage = "Ошибка при удалении файла";
+            String errorMessage = "Ошибка при удалении файла из хранилища";
             log.error(errorMessage, exception);
             throw new IntegrationException(errorMessage);
         }
@@ -64,16 +50,6 @@ public class S3Service {
             s3Client.putObject(request);
         } catch (IOException exception) {
             String errorMessage = "Ошибка при отправке файла в хранилище";
-            log.error(errorMessage, exception);
-            throw new IntegrationException(errorMessage);
-        }
-    }
-
-    public List<S3ObjectSummary> getAllObject(String folder) {
-        try {
-            return s3Client.listObjects(bucketName, folder).getObjectSummaries();
-        } catch (SdkClientException exception) {
-            String errorMessage = "Ошибка при получении файлов из хранилища";
             log.error(errorMessage, exception);
             throw new IntegrationException(errorMessage);
         }
